add: userAllUsers custom hook
