+++ src/components/custom_hook/CustomHook.tsx
@@ -1,39 +1,12 @@
-import React, { useState, VFC } from 'react';
-import axios from 'axios';
+import React, { VFC } from 'react';
 
-import { User } from '../../types/api/user';
 import UserCard from './UserCard';
-import { UserProfile } from '../../types/userProfile';
+import { useAllUsers } from './hooks/useAllUsers';
 
 const CustomHook: VFC = () => {
-  const [userProfiles, setUserProfiles] = useState<Array<UserProfile>>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(false);
+  const { getUsers, userProfiles, loading, error } = useAllUsers();
 
-  const onClickFetchData = () => {
-    setLoading(true);
-    setError(false);
-
-    axios
-      .get<Array<User>>('https://jsonplaceholder.typicode.com/users')
-      .then((res) => {
-        const data = res.data.map((user) => ({
-          id: user.id,
-          name: `${user.name}(${user.username})`,
-          email: user.email,
-          address: `${user.address.city}${user.address.suite}${user.address.street}`,
-        }));
-
-        setUserProfiles(data);
-      })
-      .catch((err) => {
-        setError(true);
-        console.error(err);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  };
+  const onClickFetchData = () => getUsers();
 
   return (
     <div>
